Core: Updated all internal referenced to the most modern versions
Detailed:
The modules in use on this project were seriously out of date, and
represented a potential security risk on websites currently using it.

I have updated all modules to their most modern versions possible,
and fixed tests within the application, alongside the build process
to ensure security is maintained.

Signed-off-by: Andrew Lowther <Andrew.Lowther@gentinguk.com>